+++ test/anyone-can-renew-test.js
@@ -163,12 +163,12 @@ describe('Anyone-can-renew address', function() {
   it('should not be owned by either wallet', async  () => {
     assert.rejects(
       alice.sendTransfer(name, aliceReceive),
-      {message: `Wallet does not own: "${name}".`}
+      {message: `Wallet does not own name: ${name}.`}
     );
 
     assert.rejects(
       bob.sendTransfer(name, bobReceive),
-      {message: 'Auction not found.'}
+      {message: `Auction not found: ${name}.`}
     );
   });